TongueDiagnosis-master/requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyio==4.3.0
boto3==1.34.59
botocore==1.34.59
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cryptography==42.0.5
cycler==0.10.0
dill==0.3.8
ecdsa==0.18.0
fastapi==0.110.0
filelock==3.13.1
fire==0.5.0
fonttools==4.49.0
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
greenlet==3.0.3
grpcio==1.62.1
h11==0.14.0
huggingface-hub==0.21.4
idna==2.10
Jinja2==3.1.3
jmespath==1.0.1
kiwisolver==1.4.5
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.8.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.7.0.72
opencv-python-headless==4.8.0.74
packaging==24.0
pandas==2.2.1
pillow==10.2.0
protobuf==4.25.3
psutil==5.9.8
py-cpuinfo==9.0.0
pyasn1==0.5.1
pybboxes==0.1.6
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
roboflow==1.1.23
rsa==4.9
s3transfer==0.10.0
sahi==0.11.15
scipy==1.12.0
seaborn==0.13.2
shapely==2.0.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
ultralytics==8.1.26
urllib3
uvicorn==0.28.0
Werkzeug==3.0.1
